Move some files around to make the flow a easier to follow
