+++ src/components/GlobalInitializer.tsx
@@ -1,0 +1,7 @@
+import { GlobalConfigParser } from './GlobalConfigParser';
+
+export function GlobalInitializer() {
+  return (
+    <GlobalConfigParser />
+  );
+}
